reverting commit faa22bcd5af1bdec09e39b4c387fc0315716ccb0 , but ordering the auto-config beans this time, so that dummy classes are instantiated only if required (they used to be instantiated first, making the @ConditionalOnMissingBean useless.)
