+++ ci-droid-tasks-consumer-autoconfigure/src/main/java/com/societegenerale/cidroid/tasks/consumer/autoconfigure/CiDroidTasksConsumerAutoConfiguration.java
@@ -13,6 +13,8 @@ import com.societegenerale.cidroid.tasks.consumer.services.RemoteGitHub;
 import com.societegenerale.cidroid.tasks.consumer.services.actionHandlers.*;
 import com.societegenerale.cidroid.tasks.consumer.services.notifiers.Notifier;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.AutoConfigureOrder;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.context.annotation.Bean;
@@ -35,26 +37,51 @@ public class CiDroidTasksConsumerAutoConfiguration {
 
     @Bean
     @ConditionalOnProperty(value = "ciDroidBehavior.notifyOwnerForNonMergeablePr.enabled", havingValue = "true")
+    @AutoConfigureOrder(1)
     public PushEventOnDefaultBranchHandler notificationHandler(RemoteGitHub gitHub, List<Notifier> notifiers) {
         return new NotificationsHandler(gitHub, notifiers);
     }
 
     @Bean
     @ConditionalOnProperty(value = "ciDroidBehavior.tryToRebaseOpenPrs.enabled", havingValue = "true")
+    @AutoConfigureOrder(2)
     public PushEventOnDefaultBranchHandler rebaseHandler(RemoteGitHub gitHub, @Value("${gitHub.login}") String gitLogin,
             @Value("${gitHub.password}") String gitPassword) {
 
         return new RebaseHandler(new GitRebaser(gitLogin, gitPassword, new GitWrapper()), gitHub);
     }
 
+
+    @Bean
+    @ConditionalOnMissingBean(PushEventOnDefaultBranchHandler.class)
+    @AutoConfigureOrder(500)
+    public PushEventOnDefaultBranchHandler dummyPushEventOnDefaultBranchHandler(){
+
+        return new DummyPushEventOnDefaultBranchHandler();
+    }
+
+
+
+
+
+
     @Bean
     @ConditionalOnProperty(value = "ciDroidBehavior.bestPracticeNotifier.enabled", havingValue = "true")
+    @AutoConfigureOrder(1)
     public PullRequestEventHandler bestPracticeNotifierHandler(CiDroidBehavior ciDroidBehavior, List<Notifier> notifiers,
             RemoteGitHub remoteGitHub) {
 
         return new BestPracticeNotifierHandler(ciDroidBehavior.getPatternToResourceMapping(), notifiers, remoteGitHub,
                 new RestTemplateResourceFetcher());
 
+    }
+
+    @Bean
+    @ConditionalOnMissingBean(PullRequestEventHandler.class)
+    @AutoConfigureOrder(500)
+    public PullRequestEventHandler dummyPullRequestEventHandler(){
+
+        return new DummyPullRequestEventHandler();
     }
 
     @Bean
