+++ ci-droid-tasks-consumer-infrastructure/src/main/java/com/societegenerale/cidroid/tasks/consumer/infrastructure/config/InfraConfig.java
@@ -7,13 +7,10 @@ import com.societegenerale.cidroid.tasks.consumer.infrastructure.FeignRemoteGitH
 import com.societegenerale.cidroid.tasks.consumer.infrastructure.GithubEventListener;
 import com.societegenerale.cidroid.tasks.consumer.infrastructure.notifiers.EMailActionNotifier;
 import com.societegenerale.cidroid.tasks.consumer.services.*;
-import com.societegenerale.cidroid.tasks.consumer.services.actionHandlers.DummyPullRequestEventHandler;
-import com.societegenerale.cidroid.tasks.consumer.services.actionHandlers.DummyPushEventOnDefaultBranchHandler;
 import com.societegenerale.cidroid.tasks.consumer.services.actionHandlers.PullRequestEventHandler;
 import com.societegenerale.cidroid.tasks.consumer.services.actionHandlers.PushEventOnDefaultBranchHandler;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.web.HttpMessageConverters;
 import org.springframework.cloud.netflix.feign.EnableFeignClients;
 import org.springframework.cloud.netflix.feign.FeignAutoConfiguration;
@@ -103,19 +100,6 @@ public class InfraConfig {
         return new PullRequestEventService(pullRequestEventHandlers);
     }
 
-    @Bean
-    @ConditionalOnMissingBean(PushEventOnDefaultBranchHandler.class)
-    public PushEventOnDefaultBranchHandler dummyPushEventOnDefaultBranchHandler(){
-
-        return new DummyPushEventOnDefaultBranchHandler();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean(PullRequestEventHandler.class)
-    public PullRequestEventHandler dummyPullRequestEventHandler(){
-
-        return new DummyPullRequestEventHandler();
-    }
 
     @Bean
     public HttpMessageConverters feignHttpMessageConverters() {
